% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{depInfo}
\alias{depInfo}
\title{Get information about external dependencies}
\usage{
depInfo()
}
\description{
Returns a \code{data.frame} with information about the external dependencies
(tools and data files) that are installed by \pkg{patRoonExt}.
}
\seealso{
\code{\link{getExtPath}}
}
