% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRoonExt.R
\name{patRoonExt}
\alias{patRoonExt}
\alias{getExtPath}
\title{External dependencies for patRoon}
\usage{
getExtPath(what, warn = TRUE)
}
\arguments{
\item{what}{The tool/data file to query, currently supported are \code{openms}, \code{sirius}, \code{openbabel}, \code{metfragcl},
\code{metfragct}, \code{metfragpcl}, \code{biotransformer}.}

\item{warn}{Throw a warning when the queried tool/file could not be found.}
}
\description{
This package bundles several tools and data files that may be of use in
\href{https://github.com/rickhelmus/patRoon}{patRoon} workflows.
}
\details{
\pkg{patRoon} uses several (non \R based) tools and data files to implement or enhance certain parts of the workflow.
This package bundles the most common to simplify their installation. Please see the \href{https://github.com/rickhelmus/patRoon/README.md}{Project's README} for more details.

The \code{getExtPath} function returns the path to a bundled tool or data file.
}
